Moved testing to the TestRun file
