+++ NeuralNetwork.py
@@ -8,6 +8,8 @@ class NeuralNetwork:
     def __init__(self, nb_features, nb_nodes_per_layer, nb_outputs, nb_hidden_layers=1, weight_decay=0.5):
         # Parameters
         self.weight_decay = weight_decay
+        self.nb_nodes_per_layer = nb_nodes_per_layer
+        self.nb_hidden_layers = nb_hidden_layers
 
         # For graphing
         self.train_error = []
@@ -51,7 +53,7 @@ class NeuralNetwork:
                     self.test_error.append(self.getCost(test_X, test_y))
 
             # TESTING decaying learning rate ###
-            if j % 100 == 0:
+            if j % 100 == 0 and learning_rate >= 0.01:
                 learning_rate *= 0.7
 
             # Initialize sum of errors
@@ -79,7 +81,6 @@ class NeuralNetwork:
                 # Output errors (deltas)
                 out_errors = out_activations - y[i]
 
-
                 # Hidden layer errors
                 hid_errors = []
                 # Hidden layer -1
@@ -90,7 +91,6 @@ class NeuralNetwork:
                 for k in reversed(range(1, len(self.hid_weights))):
                     hid_errors.insert(0, np.dot(self.hid_weights[k].T, hid_errors[0]) * self.sigmoidPrime(hid_activations[k-1]))
                     hid_errors[0] = hid_errors[0][1:]  # remove bias term
-
 
                 # Update sum of errors
                 # Hidden layer 1
@@ -104,7 +104,6 @@ class NeuralNetwork:
 
                 # Output layer
                 out_deriv_sum += np.dot(np.atleast_2d(out_errors).T, np.atleast_2d(hid_activations[-1]))
-
 
                 # in_deriv_sum += np.dot(np.atleast_2d(l1_errors_no_bias).T, np.atleast_2d(X[i]))
                 # out_deriv_sum += np.dot(np.atleast_2d(out_errors).T, np.atleast_2d(in_activations))
@@ -220,14 +219,18 @@ class NeuralNetwork:
         plt.legend()
         return
 
+    def test(self, X, y, iterations, learning_rate, test_frac=0, X_test=None, y_test=None):
+        # Trains the neural net and prints the final training and testing errors (and the minimum test error)
+        # If X_test and y_test are given then those are used as the test sets
+        # Or else test_frac is used to split the data into training and test sets
 
-    def test(self, X, y, iterations, learning_rate, test_frac):
-        # Splits the data into a training set and a test set
-        # Prints the final training error and testing error (and the minimum test error)
         n = int(len(X) * (1 - test_frac))
-        # Could shuffle examples first
 
-        self.train(X[:n], y[:n], iterations, learning_rate, X[n:], y[n:])
+        if X_test is None and y_test is None:
+            X_test = X[n:]
+            y_test = y[n:]
+
+        self.train(X[:n], y[:n], iterations, learning_rate, X_test, y_test)
 
         print("Train:", self.train_error[-1])
         print("Test:", self.test_error[-1])
@@ -236,12 +239,14 @@ class NeuralNetwork:
         print("Min:", minErr, "at", minIter, "iters")
 
         print("Train (class):", self.classError(X[:n], y[:n]))
-        print("Test (class):", self.classError(X[n:], y[n:]))
+        print("Test (class):", self.classError(X_test, y_test))
 
         return minErr, self.test_error[-1], self.train_error[-1]
 
     def testProbBuckets(self, X, y, test_frac):
         # Test probability buckets
+        # This assumes we have trained before
+
         n = int(len(X) * (1 - test_frac))
 
         preds = self.predict_mult(X)
@@ -266,88 +271,23 @@ class NeuralNetwork:
                     freq_probs_train[i] += 1
                     freq_wins_train[i] += int(y[x])
 
+        probs_test = [freq_wins_test[i]/ freq_probs_test[i] if freq_probs_test[i] != 0 else 0 for i in range(nb_buckets)]
+        probs_train = [freq_wins_train[i]/ freq_probs_train[i] if freq_probs_train[i] != 0 else 0 for i in range(nb_buckets)]
 
         print("Freq test:")
         print(freq_probs_test)
         print(freq_wins_test)
+        print(probs_test)
+
         print("Freq train:")
         print(freq_probs_train)
         print(freq_wins_train)
+        print(probs_train)
 
-        try:
-            probs_test = [freq_wins_test[i]/ freq_probs_test[i] for i in range(nb_buckets)]
-            probs_train = [freq_wins_train[i]/ freq_probs_train[i] for i in range(nb_buckets)]
-            print(probs_test)
-            print(probs_train)
-        except:
-            pass
 
     def __repr__(self):
-        return "Hidden layer weights:\n" + str(self.hid_weights) + "\nOutput layer weights:\n" + str(self.out_weights)
-
-def testRuns(n, x, y):
-    # Runs n tests and finds the average errors
-    min_errs =[]
-    final_errs = []
-    train_errs = []
-    for i in range(n):
-        net = NeuralNetwork(96, 32, 1, 2, weight_decay=25)
-        temp = net.test(x, y, 2000, 0.25, 0.3)
-
-        min_errs.append(temp[0])
-        final_errs.append(temp[1])
-        train_errs.append(temp[2])
-        net.testProbBuckets(x, y, 0.3)
-
-    print(min_errs)
-    print("Avg min:", sum(min_errs)/n)
-    print(final_errs)
-    print("Avg final test:", sum(final_errs)/n)
-    print(train_errs)
-    print("Avg final train:", sum(train_errs)/n)
-
-    return
+        return "Nb hidden layers: " + str(self.nb_hidden_layers) + "\tNb hidden nodes per layer: " + str(self.nb_nodes_per_layer) + \
+               "\tWeight decay: {0:.3f}".format(self.weight_decay)
 
 if __name__ == "__main__":
-    net = NeuralNetwork(96, 32, 1, nb_hidden_layers=2, weight_decay=25)
-
-    input = np.genfromtxt(
-    'InputData2014-15_Final.csv',           # file name
-    delimiter=',',          # column delimiter
-    dtype='float32',        # data type
-    filling_values=0,       # fill missing values with 0
-    )
-
-    random.seed()
-    random.shuffle(input)
-    x = input[:, 1:]
-    y = input[:, 0]
-
-    # net.test(x, y, 1000, 0.25, 0.3)
-    # net.graphCosts(5)
-    # net.testProbBuckets(x, y, 0.3)
-    testRuns(30, x, y)
-
-    # x = [1, 2, 3, 4, 5]
-    # print(x[3:])
-
-    # minsErr = []
-    # minsIter = []
-    #
-    # tests = [2*i for i in range(4, 20)]
-    # for j in tests:
-    #     print("--- nodes =", j)
-    #     net = NeuralNetwork(8, j, 1, weight_decay=0.4)
-    #
-    #     temp = net.test(x, y, 300, 1.5, 0.8)
-    #     minsErr.append(temp[0])
-    #     minsIter.append(temp[1])
-    #
-    # print(minsErr)
-    # print(minsIter)
-    # plt.plot(tests, minsErr, label="Error")
-    # plt.title("Error vs nodes")
-    # plt.figure()
-    # plt.plot(tests, minsIter, label="Iter")
-    # plt.title("Iter vs nodes")
-    plt.show()
+    pass
