+++ TestRun.py
@@ -1,34 +1,54 @@
 import NeuralNetwork as nn
-import csv
+import numpy as np
+import random
 import matplotlib.pyplot as plt
 
-X = []
-with open('InputData2014-15_Final.csv', 'r') as csvfile:
-    reader = csv.reader(csvfile)
-    for row in reader:
-        X.append([float(entry) for entry in row])
+# This file is used for testing the neural network
 
-csvfile.close()
+def testRuns(n, x, y):
+    # Runs n tests and finds the average errors
+    min_errs =[]
+    final_errs = []
+    train_errs = []
+    for i in range(n):
+        print("--- Run " + str(i+1) + " ---")
+        net = nn.NeuralNetwork(96, 32, 1, 2, weight_decay=25)
+        temp = net.test(x, y, 2000, 0.25, 0.3)
+
+        min_errs.append(temp[0])
+        final_errs.append(temp[1])
+        train_errs.append(temp[2])
+        net.testProbBuckets(x, y, 0.3)
+
+    print(min_errs)
+    print("Avg min:", sum(min_errs)/n)
+    print(final_errs)
+    print("Avg final test:", sum(final_errs)/n)
+    print(train_errs)
+    print("Avg final train:", sum(train_errs)/n)
+
+    return
+
+if __name__ == '__main__':
+
+    net = nn.NeuralNetwork(96, 32, 1, nb_hidden_layers=2, weight_decay=25)
+
+    input = np.genfromtxt(
+    'InputData2014-15_Final.csv',           # file name
+    delimiter=',',          # column delimiter
+    dtype='float32',        # data type
+    filling_values=0,       # fill missing values with 0
+    )
+
+    random.seed()
+    random.shuffle(input)
+    x = input[:, 1:]
+    y = input[:, 0]
 
 
-y = []
-with open('GameData2014-15_Clean.csv', 'r') as csvfile:
-    reader = csv.reader(csvfile)
-    for row in reader:
-        y.append([float(entry) for entry in row])
+    net.train(x, y, 200, 0.04, True)
 
-csvfile.close()
+    net.graphCosts()
 
-for i in range(4):
-    print(X[i], y[i])
-
-net = nn.NeuralNetwork(5, 47, 2)
-
-net.train(X, y, 200, 0.04, True)
-
-
-print(net.predict(X[0], y[0]))
-net.graphCosts()
-
-plt.show()
+    plt.show()
